backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
websockets==12.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
supabase==2.3.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
